Move setup to index
